requirements.txt
matplotlib>=3.0.0
numpy>=1.20.0    # matplotlib dependency
streamlit>=1.0.0
librosa==0.10.0
soundfile==0.12.1
setuptools>=58.0.0
google-generativeai
python-dotenv